Check miniopath exists (#13364)

+++ tools/cluster-management/utils/snapshot-utils.sh
@@ -468,7 +468,12 @@ function cleanupBackupStorage() {
 
       log "Cleaning up wal files in minio bucket ${minioBucket}. Will delete all files at path ${pathToDelete}"
       doContinue
-      kubectl_common exec "${minioPod}" -- mc rm --recursive --force "${pathToDelete}"
+
+      if kubectl_common exec "${minioPod}" -- mc stat "${pathToDelete}" >/dev/null 2>&1; then
+        kubectl_common exec "${minioPod}" -- mc rm --recursive --force "${pathToDelete}"
+      else
+        log "Skipping cleanup. Path does not exist: ${pathToDelete}"
+      fi
     done
   fi
 }
